Refactor: Utils move to utils package
