+++ src/main/groovy/org/jenkinsci/plugins/yamlaxis/YamlAxis.groovy
@@ -8,6 +8,7 @@ import hudson.matrix.AxisDescriptor
 import hudson.matrix.MatrixBuild
 import hudson.util.FormValidation
 import net.sf.json.JSONObject
+import org.jenkinsci.plugins.yamlaxis.util.BuildUtils
 import org.kohsuke.stapler.DataBoundConstructor
 import org.kohsuke.stapler.QueryParameter
 import org.kohsuke.stapler.StaplerRequest
